+++ package.json
@@ -25,6 +25,7 @@
     "@types/bcryptjs": "^2.4.6",
     "axios": "^1.11.0",
     "bcryptjs": "^3.0.2",
+    "canvas-confetti": "^1.9.3",
     "framer-motion": "^12.23.12",
     "katex": "^0.16.22",
     "mathjs": "^14.6.0",
@@ -48,6 +49,7 @@
     "@tanstack/eslint-plugin-query": "^5.83.1",
     "@testing-library/jest-dom": "^6.6.4",
     "@testing-library/react": "^16.3.0",
+    "@types/canvas-confetti": "^1.9.0",
     "@types/jest": "^30.0.0",
     "@types/node": "^20",
     "@types/react": "^19",
