+++ utils/logs/logos.ts
@@ -1,0 +1,9 @@
+export const logos = [
+  { src: '/logos/logo-Sr.svg', alt: 'Sr' },
+  { src: '/logos/logo-Sy.svg', alt: 'Sy' },
+  { src: '/logos/logo-Eb.svg', alt: 'Eb' },
+  { src: '/logos/logo-M.svg', alt: 'M' },
+  { src: '/logos/logo-T.svg', alt: 'T' },
+  { src: '/logos/logo-L.svg', alt: 'L' },
+  { src: '/logos/logo-Eg.svg', alt: 'Eg' },
+];
